Changed all BigDecimals back to double.

+++ client/lib/api/UserPlant.dart
@@ -8,7 +8,7 @@ class UserPlant implements JsonSerializeable {
   int id;
   int userId;
   String nickname;
-  double potVolume;
+  int potVolume;
   double latitude;
   double longitude;
   String imageName;
